requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
faiss-gpu
torch
torchvision
torchaudio
transformers
requests
beautifulsoup4
tqdm
pdfplumber
pymupdf
PyPDF2
pdf2image
pytesseract
easyocr
selenium
webdriver-manager
concurrent-log-handler